refactor(apps): rename app mount export project -> mount
The factory each app ships returns a Mount and is default-exported by an
epicenter.config.ts, but it lived at `@epicenter/<app>/project`. "project" was
a synonym for the one concept the whole CLI/daemon surface already names
(Mount, defineMount, MountContext, the <mount>.<action> prefix), and it
collides with the universal "project = repo/package" meaning, so the import
name no longer pointed at the value it returns.

Rename the file (project.ts -> mount.ts) and the subpath export (./project ->
./mount) across fuji, honeycrisp, opensidian, zhongwen, and tab-manager, and
update the lone code importer (examples/fuji/epicenter.config.ts), the init
scaffold default, and the module-map JSDoc and READMEs. Now the import name,
the Mount type, and the CLI prefix are one word.
