+++ apps/fuji/package.json
@@ -6,7 +6,7 @@
 	"type": "module",
 	"exports": {
 		".": "./src/lib/workspace/index.ts",
-		"./project": "./project.ts"
+		"./mount": "./mount.ts"
 	},
 	"imports": {
 		"#platform/auth": {
