--- apps/fuji/project.ts
@@ -1,8 +1,0 @@
-/**
- * Fuji project mount entry point.
- *
- * Keep this file at the app root so local project configs can import
- * `../epicenter/apps/fuji/project.ts`, matching the other app packages.
- */
-
-export { fuji } from './src/lib/workspace/project.js';
