+++ apps/opensidian/package.json
@@ -7,7 +7,7 @@
 	"exports": {
 		".": "./opensidian.ts",
 		"./browser": "./opensidian.browser.ts",
-		"./project": "./project.ts"
+		"./mount": "./mount.ts"
 	},
 	"scripts": {
 		"dev": "bun run dev:local",
