+++ apps/tab-manager/mount.ts
@@ -1,11 +1,9 @@
 /**
- * Honeycrisp project mount.
+ * Tab Manager mount.
  *
- * `honeycrisp(opts?)` returns the `Mount` that an
- * `epicenter.config.ts` default-exports. Disk paths follow the
- * Epicenter-root layout: the SQLite mirror at `.epicenter/sqlite/<id>.db`
- * (hidden) and the read-only markdown projection under table-named folders in
- * the app root.
+ * `tabManager(opts?)` returns the Mount used by `epicenter.config.ts`.
+ * It projects saved tabs, bookmarks, and devices into markdown while keeping
+ * the Y.Doc update log and SQLite mirror under `.epicenter/`.
  */
 
 import { join } from 'node:path';
@@ -22,9 +20,10 @@ import {
 	sqlitePath,
 } from '@epicenter/workspace/node';
 import { createLogger } from 'wellcrafted/logger';
-import { honeycrispWorkspace } from './honeycrisp.js';
+import { createTabManager } from './src/lib/workspace/definition.js';
 
-export type HoneycrispMountOptions = {
+export type TabManagerMountOptions = {
+	/** Enable per-materializer Git autosave for markdown output. */
 	git?: GitAutosaveConfig;
 	/**
 	 * Base URL of the Epicenter cloud API used for sync.
@@ -33,9 +32,9 @@ export type HoneycrispMountOptions = {
 	baseURL?: string;
 };
 
-export function honeycrisp(opts: HoneycrispMountOptions = {}) {
+export function tabManager(opts: TabManagerMountOptions = {}) {
 	return defineSessionMount({
-		name: 'honeycrisp',
+		name: 'tab-manager',
 		open(ctx) {
 			const { epicenterRoot, mount } = ctx;
 			const baseURL =
@@ -43,27 +42,35 @@ export function honeycrisp(opts: HoneycrispMountOptions = {}) {
 				process.env.EPICENTER_API_URL ||
 				'https://api.epicenter.so';
 
-			const workspace = honeycrispWorkspace.create();
+			const workspace = createTabManager();
 
 			const sqlite = attachBunSqliteMaterializer(workspace, {
 				filePath: sqlitePath(epicenterRoot, workspace.ydoc.guid),
+				fts: {
+					bookmarks: ['title', 'url'],
+					savedTabs: ['title', 'url'],
+				},
 				log: createLogger(`${mount}-sqlite`),
 			});
-
 			const markdown = attachMarkdownExport(workspace, {
 				dir: epicenterRoot,
-				tables: { notes: {} },
+				tables: {
+					bookmarks: {},
+					devices: {},
+					savedTabs: {},
+				},
 			});
 			if (opts.git) {
-				attachGitAutosave({
-					ydoc: workspace.ydoc,
-					dir: join(epicenterRoot, 'notes'),
-					config: opts.git,
-				});
+				for (const tableDir of ['bookmarks', 'devices', 'savedTabs']) {
+					attachGitAutosave({
+						ydoc: workspace.ydoc,
+						dir: join(epicenterRoot, tableDir),
+						config: opts.git,
+					});
+				}
 			}
 
 			const actions = defineActions({
-				...workspace.actions,
 				...sqlite.actions,
 				...markdown.actions,
 			});
